refactor: pane 生成オーケストレーションを internal/app/panelaunch へ統合 (layered 8/10)
cmd/fanout/pane.go(812 行)と tui_launch.go に重複していた 3 本の
生成列(createPaneDetailed / createAttachedPane / launchShellPane)を
internal/app/panelaunch の Launcher(Launch/LaunchOK/Attach/Shell)に
統合する。tmux split → 装飾 → relayout の共有部は splitAndDecorate、
失敗時の掃除は failCleanup に一本化(strict/best-effort の差は
decorateOpts で表現)。Request ビルダーは 7 種 → 5 コンストラクタに整理。

挙動不変(検証エージェントが旧関数との行単位 diff で確認):
- RefreshWarning → prepErr、relayout → codex WaitReady、
  rollbackState → cleanup の順序維持
- Attach は WorktreeCreated hook を叩かない / Shell は全装飾 best-effort
- dry-run 出力・shellQuote は verbatim(Tier 2 golden バイト不変)

pane_test.go(765 行)は panelaunch の白箱テストへ移送。
