+++ cmd/fanout/plan_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	"github.com/butaosuinu/fanout/internal/app/cliflags"
+	"github.com/butaosuinu/fanout/internal/app/panelaunch"
 	"github.com/butaosuinu/fanout/internal/infra/ghissue"
 )
 
@@ -134,7 +135,7 @@ func TestExistingWorktreeFannedPreservesCurrentParentQualifiedFallback(t *testin
 
 func TestExistingWorktreeFannedIgnoresGeneratedManualPaneSlug(t *testing.T) {
 	root := t.TempDir()
-	mkdirAll(t, filepath.Join(root, ".fanout", "worktrees", manualPaneSlug("Diagnostics", -12)))
+	mkdirAll(t, filepath.Join(root, ".fanout", "worktrees", panelaunch.ManualPaneSlug("Diagnostics", -12)))
 	cfg := &cliflags.Config{ParentRef: "200"}
 	issues := []ghissue.Issue{{Number: 12, Title: "Diagnostics", State: "OPEN"}}
