+++ cmd/fanout/tui_restore.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/butaosuinu/fanout/internal/app/panelaunch"
 	"github.com/butaosuinu/fanout/internal/app/panelayout"
 	"github.com/butaosuinu/fanout/internal/core/agent"
 	"github.com/butaosuinu/fanout/internal/infra/codexapp"
@@ -263,13 +264,13 @@ func recreateRecordedPane(pane state.Pane, root, session, commandName string) (s
 			return pane, fmt.Errorf("restore pane liveness key on %s: %w", paneID, err)
 		}
 	}
-	_ = tmuxrun.SetPaneTitle(paneID, restorePaneTitle(pane))                           // cosmetic; pane is still usable if tmux rejects title updates
-	_ = tmuxrun.SetPaneLabel(paneID, borderLabel(pane.Parent, restorePaneTitle(pane))) // cosmetic pane-border label
-	_ = tmuxrun.EnablePaneBorderTitles(paneID)                                         // cosmetic pane-border label
-	_ = tmuxrun.SetPaneProjectRoot(paneID, root)                                       // best-effort dashboard keybinding hint
-	_ = tmuxrun.SetPaneWorktreePath(paneID, pane.WorktreePath)                         // best-effort same-worktree action target
+	_ = tmuxrun.SetPaneTitle(paneID, restorePaneTitle(pane))                                      // cosmetic; pane is still usable if tmux rejects title updates
+	_ = tmuxrun.SetPaneLabel(paneID, panelaunch.BorderLabel(pane.Parent, restorePaneTitle(pane))) // cosmetic pane-border label
+	_ = tmuxrun.EnablePaneBorderTitles(paneID)                                                    // cosmetic pane-border label
+	_ = tmuxrun.SetPaneProjectRoot(paneID, root)                                                  // best-effort dashboard keybinding hint
+	_ = tmuxrun.SetPaneWorktreePath(paneID, pane.WorktreePath)                                    // best-effort same-worktree action target
 	if statusPath != "" {
-		status, err := codexapp.WaitReady(statusPath, codexPlanTUIStartupTimeout)
+		status, err := codexapp.WaitReady(statusPath, panelaunch.CodexPlanTUIStartupTimeout)
 		_ = os.Remove(statusPath)
 		if err != nil {
 			_ = tmuxrun.KillPane(paneID)
