+++ internal/app/panelaunch/shell.go
@@ -1,0 +1,164 @@
+package panelaunch
+
+import (
+	"crypto/rand"
+	"encoding/hex"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+	"time"
+
+	"github.com/butaosuinu/fanout/internal/app/panelayout"
+	"github.com/butaosuinu/fanout/internal/infra/state"
+	"github.com/butaosuinu/fanout/internal/infra/tmuxrun"
+	"github.com/butaosuinu/fanout/internal/infra/worktree"
+)
+
+// ShellRequest describes one plain terminal pane to open.
+type ShellRequest struct {
+	// TargetPath is the directory the shell starts in.
+	TargetPath string
+	// Root marks the project-root terminal (names the pane "root terminal").
+	Root bool
+}
+
+// Shell opens a plain shell pane at req.TargetPath and records it as a
+// keyed @manual shell row in l.Info.ProjectRoot's state. It locks that
+// project's state itself; only l.Info is used.
+func (l *Launcher) Shell(req ShellRequest) error {
+	projectRoot := l.Info.ProjectRoot
+	target := l.Info.Target
+	rawPath := strings.TrimSpace(req.TargetPath)
+	if rawPath == "" {
+		return fmt.Errorf("terminal path is required")
+	}
+	targetPath, err := filepath.Abs(rawPath)
+	if err != nil {
+		return fmt.Errorf("resolve terminal path: %w", err)
+	}
+	st, statErr := os.Stat(targetPath)
+	if statErr != nil {
+		return fmt.Errorf("terminal path: %w", statErr)
+	} else if !st.IsDir() {
+		return fmt.Errorf("terminal path is not a directory: %s", targetPath)
+	}
+
+	if excludeErr := worktree.EnsureLocalExclude(projectRoot); excludeErr != nil {
+		return fmt.Errorf("prepare local git exclude: %w", excludeErr)
+	}
+
+	recorder, err := state.LockProject(projectRoot)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		_ = recorder.Unlock()
+	}()
+
+	number := NextSyntheticPaneNumber(recorder.Store, ManualParentRef)
+	slug := shellPaneSlug(targetPath, req.Root, number)
+	title := shellPaneTitle(targetPath, req.Root)
+	shellKey, err := NewShellPaneKey()
+	if err != nil {
+		return err
+	}
+	paneID, err := tmuxrun.SplitPane(target, targetPath)
+	if err != nil {
+		return err
+	}
+	if err := tmuxrun.SetPaneShellKey(paneID, shellKey); err != nil {
+		_ = tmuxrun.KillPane(paneID)
+		return err
+	}
+	// Shell pane ergonomics are best-effort; the recorded pane id is enough to
+	// keep the terminal usable when tmux metadata/layout updates fail.
+	_ = tmuxrun.SetPaneTitle(paneID, title)
+	_ = tmuxrun.SetPaneLabel(paneID, BorderLabel(ManualParentRef, title))
+	_ = tmuxrun.EnablePaneBorderTitles(paneID)
+	_ = tmuxrun.SetPaneProjectRoot(paneID, projectRoot)
+	_ = tmuxrun.SetPaneWorktreePath(paneID, targetPath)
+	if err := recorder.RecordPane(state.Pane{
+		Parent:       ManualParentRef,
+		IssueNum:     number,
+		Kind:         state.PaneKindShell,
+		Slug:         slug,
+		PaneID:       paneID,
+		ShellKey:     shellKey,
+		Agent:        state.PaneKindShell,
+		DisplayName:  title,
+		WorktreePath: targetPath,
+		CreatedAt:    time.Now().UTC().Format(time.RFC3339),
+	}); err != nil {
+		_ = tmuxrun.KillPane(paneID)
+		// Reconcile any spacer a concurrent resize relayout may have created for
+		// this now-killed pane, so no blank pane is left behind.
+		_ = panelayout.Apply(target, panelayout.Close)
+		return fmt.Errorf("write fanout state: %w", err)
+	}
+	// Re-layout only after the pane is recorded, so a failed/rolled-back launch
+	// never leaves the window arranged around a pane that no longer exists or an
+	// orphaned spacer behind.
+	_ = panelayout.Apply(target, panelayout.Create)
+	return nil
+}
+
+// NewShellPaneKey generates a random @fanout_shell_key liveness token.
+func NewShellPaneKey() (string, error) {
+	var b [16]byte
+	if _, err := rand.Read(b[:]); err != nil {
+		return "", fmt.Errorf("generate terminal identity: %w", err)
+	}
+	return "shell-" + hex.EncodeToString(b[:]), nil
+}
+
+func shellPaneSlug(targetPath string, root bool, number int) string {
+	base := "root"
+	if !root {
+		base = SanitizeSessionPart(filepath.Base(targetPath))
+	}
+	if base == "" {
+		base = "terminal"
+	}
+	n := number
+	if n < 0 {
+		n = -n
+	}
+	return fmt.Sprintf("terminal-%s-%d", base, n)
+}
+
+func shellPaneTitle(targetPath string, root bool) string {
+	if root {
+		return "root terminal"
+	}
+	base := strings.TrimSpace(filepath.Base(targetPath))
+	if base == "" || base == "." || base == string(filepath.Separator) {
+		base = "worktree"
+	}
+	return "terminal " + base
+}
+
+// SanitizeSessionPart mirrors cmd/fanout's tmux session-name sanitizer
+// (cmd/fanout/tui.go): lowercase [a-z0-9] runs joined by single dashes.
+func SanitizeSessionPart(s string) string {
+	s = strings.ToLower(strings.TrimSpace(s))
+	var b strings.Builder
+	lastDash := false
+	for _, r := range s {
+		allowed := (r >= 'a' && r <= 'z') || (r >= '0' && r <= '9')
+		if allowed {
+			b.WriteRune(r)
+			lastDash = false
+			continue
+		}
+		if !lastDash {
+			b.WriteByte('-')
+			lastDash = true
+		}
+	}
+	out := strings.Trim(b.String(), "-")
+	if out == "" {
+		return "repo"
+	}
+	return out
+}
